src/main/java/org/meanbean/factories/basic/LongFactory.java
package org.meanbean.factories.basic;

import org.meanbean.util.RandomNumberGenerator;

/**
 * Concrete Factory that creates Long objects.
 * 
 * @author Graham Williamson
 */
public final class LongFactory extends FactoryBase<Long> {

    /** Unique version ID of this Serializable class. */
    private static final long serialVersionUID = 1L;
    
    /**
     * Construct a new Integer object factory.
     * 
     * @param randomNumberGenerator
     *            A random number generator used by the Factory to generate random values.
     *            
     * @throws IllegalArgumentException
     *             If the specified randomNumberGenerator is deemed illegal. For example, if it is null.
     */
    public LongFactory(RandomNumberGenerator randomNumberGenerator) throws IllegalArgumentException {
        super(randomNumberGenerator);
    }

    /**
     * Create a new Long object.
     * 
     * @return A new Long object.
     */
    @Override
    public Long create() {
        return getRandomNumberGenerator().nextLong();
    }
}